Features.vue update
